dynamic search functionality concept. needs some more tlc
