+++ src/components/component-scripts/SearchFilter.js
@@ -1,20 +1,27 @@
-const axios = require('axios');
-export default{
-    data(){
-        return {
-            searchQuery: '',
-            items: []
-        };
-    },
+// const axios = require('axios');
+// export default{
+//     // data(){
+//     //     return {
+//     //         searchQuery: '',
+//     //         items: []
+//     //     };
+//     // },
 
-    computed: {
-        filteredItems(){
+//     // computed: {
+//     //     filteredItems: function(){
+//     //         return this.items.filter((item) => {
+//     //             return item.name.match(this.searchQuery)
+//     //         });
+            
+//     //     }
+//     // },
+//     // async mounted(){
+//     //     const response = await axios.get('http://ec2-54-167-36-58.compute-1.amazonaws.com:3000/item');
+//     //     this.items = response.data;
+//     //     console.log(this.items);
+//     // },
 
-        }
-    },
-    async mounted(){
-        const response = await axios.get('http://ec2-54-167-36-58.compute-1.amazonaws.com:3000/item');
-        this.items = response.data;
-        console.log(this.items);
-    }
-}
+//     // async getData(){
+//     //     console.log(this.searchQuery)
+//     // }
+// }
