+++ src/components/component-templates/SearchFilter.html
@@ -10,8 +10,8 @@
     
         <div class="container-fluid SearchFilter_container">
             <div class="row align-items-center">
-                <div class="row list-entry" v-show="toggle" v-else v-for="(searchFilter, i) in this.$store.state.searchFilter">
-                    <div class="col-2">{{ searchFilter.name }}</div>
+                <div class="row list-entry" v-for="item in filteredItems">
+                    <div class="col-2">{{ item.name }}</div>
                 </div>
             </div>
         </div>
